requirements.txt
regex==2021.4.4
altair==4.1.0
gsheetsdb==0.1.13
plotly==5.3.1
pytz==2021.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
openpyxl==3.0.7
Pillow==8.3.2
protobuf==3.17.3
scikit_learn==0.24.2
